fix(#490): make root / page auth-aware for authenticated users
Extract DashboardShell from dashboard layout into a reusable component.
Root page now conditionally renders the dashboard shell for logged-in
users and the guest public directory for visitors.
