+++ quotevote-frontend/src/app/components/AuthAwareHome.tsx
@@ -1,0 +1,51 @@
+'use client'
+
+import { type ReactElement } from 'react'
+import { useSearchParams } from 'next/navigation'
+import { useAppStore } from '@/store'
+import { DashboardShell } from '@/components/DashboardShell'
+import { PublicDirectoryContent } from './PublicDirectory/PublicDirectoryContent'
+import { DirectoryToolbar } from './PublicDirectory/DirectoryToolbar'
+import PaginatedPostsList from '@/components/Post/PaginatedPostsList'
+
+export function AuthAwareHome(): ReactElement {
+  const user = useAppStore((s) => s.user.data)
+  const loggedIn = !!(user?.id || user?._id)
+
+  if (!loggedIn) {
+    return <PublicDirectoryContent />
+  }
+
+  return <AuthenticatedFeed />
+}
+
+function AuthenticatedFeed(): ReactElement {
+  const searchParams = useSearchParams()
+
+  const q = searchParams.get('q') || ''
+  const from = searchParams.get('from') || ''
+  const to = searchParams.get('to') || ''
+  const sortParam = searchParams.get('sort')
+  const sortOrder = sortParam === 'asc' ? 'asc' : 'desc'
+  const interactions = searchParams.get('interactions') === 'true'
+  const groupId = searchParams.get('group') || undefined
+
+  return (
+    <DashboardShell>
+      <DirectoryToolbar />
+      <div className="w-full max-w-2xl mx-auto min-w-0">
+        <PaginatedPostsList
+          defaultPageSize={20}
+          maxVisiblePages={5}
+          searchKey={q}
+          startDateRange={from || undefined}
+          endDateRange={to || undefined}
+          sortOrder={sortOrder}
+          interactions={interactions}
+          groupId={groupId}
+          compact
+        />
+      </div>
+    </DashboardShell>
+  )
+}
